+++ Changelog.md
@@ -4,6 +4,7 @@ Lilu Changelog
 - Fixed AZAL recognition as GPU audio on certain AMD platforms (thx to wkpark)
 - Added external GPU disabling API with device and kernel selection via properties
 - Added identifiers for Rocket Lake and Tiger Lake CPUs
+- Added API to disable builtin GPU (IGPU)
 
 #### v1.5.1
 - Added `lilu_os_memmem` and `lilu_os_memchr` APIs